Fix Create User Invited Issue
