+++ pages/freelancers/index.js
@@ -191,7 +191,7 @@ const Freelancers = ({
         item?.user?.profileImage || 'https://res.cloudinary.com/dghsmwkfq/image/upload/v1670086178/dinosaur_xzmzq3.png',
       rate: item?.rate,
       likes: item?.likeTotal,
-      isInvited: item?.invites && item?.invites.userInvited == user ? true : false
+      isInvited: item?.invites && item?.invites.userInvited == userId ? true : false
     }
     return freelancer
   }
